symlink as file

+++ app/src/java/pd/droidapp/fmgr/util/FileProperties.java
@@ -13,7 +13,7 @@ public class FileProperties {
     public Long size;
     public String sha256sum;
 
-    public Integer numChildren;
+    public Long numChildren;
 
     public FileProperties(String path, boolean isDirectory) {
         this.path = path;
